// src/lib/generated/cups.generated.ts
export type CupsModality = "RX" | "TAC" | "ECO" | "RMN" | "MG" | "DENSITO" | "HEMODINAMIA";